% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{chapter_templates}
\alias{chapter_templates}
\title{Chapter templates shipped with the package}
\usage{
chapter_templates()
}
\value{
A character vector of template names accepted by the \code{template}
argument of \code{\link[=add_chapter]{add_chapter()}}.
}
\description{
Chapter templates shipped with the package
}
\examples{
chapter_templates()
}
