% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{publication_root}
\alias{publication_root}
\title{Locate the root of a publication project}
\usage{
publication_root(path = ".")
}
\arguments{
\item{path}{A path inside a publication project. Defaults to the working
directory.}
}
\value{
The absolute path to the project root, as a character string.
}
\description{
Walks up from \code{path} looking for a directory that contains \verb{_bookdown.yml}.
}
\examples{
\dontrun{
publication_root()
}
}
